Add keeper reward plus coll distributor (#1994)
* wip

* Introduce collateralLiquidateRewardRatioD18

* add reward distributor

* wip

* wip

* duh... make it build again

* some fixes

* remove storage checking on distributor auxiliar

* Introduce CollateralConfigurationModule

* include in cannonfile

* move to collateralConfigurationModule

* comments and TODOs

* add some comments

* wip

* add collateral rewards distributor to bootstrap

* Add version to PerpsRewardDistributor

* some fixes while adding tests

* wip test

* wip compilation fixes after rebase

* make it compile

* (wip) fix tests

* lint

* update old dependencies

---------

+++ markets/perps-market/contracts/storage/GlobalPerpsMarket.sol
@@ -155,7 +155,7 @@ library GlobalPerpsMarket {
     ) internal view {
         uint256 collateralAmount = self.collateralAmounts[synthMarketId];
         if (synthAmount > 0) {
-            uint maxAmount = CollateralConfiguration.load(synthMarketId).maxAmount;
+            uint256 maxAmount = CollateralConfiguration.load(synthMarketId).maxAmount;
             if (maxAmount == 0) {
                 revert SynthNotEnabledForCollateral(synthMarketId);
             }
